Finished creating test world state from json
Also creates the relationships as defined in the "relationships" object
in the input json.

This means we have all the basic functionality needed to create a graph
structure out of the entities and relationships extracted by the LLM

need to create a better means of defining the relationship type,
currently need to write a separate query for each relationship type.
Also need to test what happens if attempt to create a node which already
exists. This should work due to using MERGE but this is untested. It
should also not overwrite the existing node but do something smart with
description e.g. append further description on top of what exists