Fix find-k-similar-descriptions, return only k
previously this was returning some large number of results which were
all duplicates of the first k that was requested.
Not sure if this is/was some bug with the bolt driver or with the
clojure interface to the bolt driver. When i ran the query directly in
the neo4j console it returned k nodes as expected.